create helper utils for gatsby-theme-series
